requirements.txt
# Core data and ML libraries
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
yfinance==0.2.18
alpha-vantage==2.3.1

# Trading and financial libraries
ta-lib==0.4.28
stockstats==0.6.2
backtrader==1.9.78.123

# Web scraping and news analysis
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
textblob==0.17.1
vaderSentiment==3.3.2

# Web dashboard
streamlit==1.29.0
plotly==5.17.0
dash==2.16.1

# Database and storage
sqlite3
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
APScheduler==3.10.4
python-telegram-bot==20.7

# API and async
aiohttp==3.9.1
asyncio

# Logging and monitoring
loguru==0.7.2 